combine tools into one
